fix: moving access to folder core
