+++ src/domains/core/access/tests/access.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing'
-import { AccessService } from './access.service'
-import { AccessController } from './access.controller'
-import { AccessFixture, CreateAccessFixture, UpdateAccessFixture } from './fixtures'
+import { AccessService } from '../access.service'
+import { AccessController } from '../access.controller'
+import { AccessFixture, CreateAccessFixture, UpdateAccessFixture } from '../access.fixtures'
 import { DeleteResult } from 'typeorm'
 
 describe('AccessController', () => {
